moving EmptyCAR test to nodebuilder/share
